+++ Homeworks/S2Hw1Zoo/Abstract/Wild.java
@@ -4,7 +4,7 @@ public abstract class Wild extends Animal{
     protected String habitat;
     protected String foundDate;
 
-    public Wild(int height, Double weight, String eyesColor, String habitat, String foundDate) {
+    public Wild(int height, int weight, String eyesColor, String habitat, String foundDate) {
         super(height, weight, eyesColor);
         this.habitat = habitat;
         this.foundDate = foundDate;
