+++ Homeworks/S2Hw1Zoo/WildAnimals/Tiger.java
@@ -4,7 +4,7 @@ import Homeworks.S2Hw1Zoo.Abstract.Wild;
 
 public class Tiger extends Wild {
 
-    public Tiger(int height, Double weight, String eyesColor, String habitat, String foundDate) {
+    public Tiger(int height, int weight, String eyesColor, String habitat, String foundDate) {
         super(height, weight, eyesColor, habitat, foundDate);
     }
