+++ Homeworks/S2Hw1Zoo/WildAnimals/Wolf.java
@@ -6,7 +6,7 @@ import Homeworks.S2Hw1Zoo.Interfaces.IpackLeader;
 public class Wolf extends Wild implements IpackLeader {
     protected boolean packLeader;
 
-    public Wolf(int height, Double weight, String eyesColor, String habitat, String foundDate, boolean packLeader) {
+    public Wolf(int height, int weight, String eyesColor, String habitat, String foundDate, boolean packLeader) {
         super(height, weight, eyesColor, habitat, foundDate);
         this.packLeader = packLeader;
     }
